ansible/requirements.txt
boto3
botocore
psutil
pywinrm
